split out UI components
